delete non existing root folders

+++ www/modules/files/controller/FolderController.php
@@ -83,28 +83,47 @@ class GO_Files_Controller_Folder extends GO_Base_Controller_AbstractModelControl
 
 		echo "Done\n";
 
-		if(!isset($params['path'])){
-			GO_Base_Fs_File::setAllowDeletes($oldAllowDeletes);
-			$folders = array('email', 'billing/notifications');
+		if(!empty($params['delete'])){
+			if(!isset($params['path'])){
+				GO_Base_Fs_File::setAllowDeletes($oldAllowDeletes);
+				$folders = array('email', 'billing/notifications');
 
-			foreach($folders as $name){
+				foreach($folders as $name){
 
-				echo "Deleting ".$name."\n";
-				GO_Files_Model_Folder::$deleteInDatabaseOnly=true;
-				GO_Files_Model_File::$deleteInDatabaseOnly=true;
-				try{
-					$folder = GO_Files_Model_Folder::model()->findByPath($name);
-					if($folder)
-							$folder->delete();
-				}
-				catch(Exception $e){
-					if (PHP_SAPI != 'cli')
-						echo "<span style='color:red;'>".$e->getMessage()."</span>\n";
-					else
-						echo $e->getMessage()."\n";
+					echo "Deleting ".$name."\n";
+					GO_Files_Model_Folder::$deleteInDatabaseOnly=true;
+					GO_Files_Model_File::$deleteInDatabaseOnly=true;
+					try{
+						$folder = GO_Files_Model_Folder::model()->findByPath($name);
+						if($folder)
+								$folder->delete();
+					}
+					catch(Exception $e){
+						if (PHP_SAPI != 'cli')
+							echo "<span style='color:red;'>".$e->getMessage()."</span>\n";
+						else
+							echo $e->getMessage()."\n";
+					}
 				}
 			}
+
+
+			$findParams = GO_Base_Db_FindParams::newInstance();
+
+			$findParams->getCriteria()->addCondition('parent_id', null,'IS');
+
+			$stmt = GO_Files_Model_Folder::model()->find($findParams);
+
+			foreach($stmt as $folder){
+
+				if(!$folder->fsFolder->exists()){
+
+					echo "Deleting ".$folder->path()."\n";
+				}
+
+			}
 		}
+		
 	}
 
 	private function _getExpandFolderIds($params){
